Recreating terminal
