+++ Echo.TerminalOld/src/CommandsController.cs
@@ -1,12 +1,15 @@
+using System;
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Globalization;
+using System.Linq;
 using System.Reflection;
 using CodeHelpers;
 using CodeHelpers.Collections;
 using CodeHelpers.Diagnostics;
 using CodeHelpers.Packed;
 
-namespace Echo.Terminal;
+namespace Echo.TerminalOld;
 
 public class CommandsController : Terminal.Section, ILogger
 {
